LayoutTests/platform/ios-sim/fast/forms/checkValidity-001-expected.txt

This test checks if checkValidity() returns correctly a true (meaning no error) result.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


 lorem ipsum    
PASS document.getElementsByTagName("fieldset")[0].checkValidity() is true
PASS document.getElementsByTagName("input")[0].checkValidity() is true
PASS document.getElementsByTagName("button")[0].checkValidity() is true
PASS document.getElementsByTagName("select")[0].checkValidity() is true
PASS document.getElementsByTagName("textarea")[0].checkValidity() is true
PASS document.getElementsByTagName("output")[0].checkValidity() is true
PASS document.getElementsByTagName("object")[0].checkValidity() is true
PASS document.getElementsByTagName("keygen")[0].checkValidity() is true
PASS successfullyParsed is true

TEST COMPLETE
